Refactor users route and separate controllers
