+++ vidly/routes/users.js
@@ -1,31 +1,11 @@
-const _ = require('lodash');
-const bcrypt = require('bcrypt');
-const express = require('express');
+const { Router } = require('express');
 const auth = require('../middleware/auth');
-const { User, validate } = require('../models/user');
+const { show, create } = require('../controllers/users');
 
-const router = express.Router();
+const router = Router();
 
-router.get('/me', auth, async (req, res) => {
-  const user = await User.findById(req.user._id);
-  return res.send(_.pick(user, ['name', 'email']));
-});
+router.get('/me', auth, show);
 
-router.post('/', async (req, res) => {
-  const { error } = validate(req.body);
-  if (error) return res.status(400).send(error.details[0].message);
-
-  let user = await User.findOne({ email: req.body.email });
-
-  if (user) return res.status(400).send('User already exists!');
-
-  user = new User(_.pick(req.body, ['name', 'email', 'password']));
-  const salt = await bcrypt.genSalt(10);
-  user.password = await bcrypt.hash(user.password, salt);
-
-  await user.save();
-  const token = user.generateAuthToken();
-  return res.header('x-auth-token', token).send(_.pick(user, ['name', 'email']));
-});
+router.post('/', create);
 
 module.exports = router;
